Coarsely split the EC2 integration test in two
