+++ package.json
@@ -50,6 +50,7 @@
     "@types/lodash.isequal": "^4.5.6",
     "@types/node": "^16.11.28",
     "@types/pg": "^8.6.5",
+    "@types/sshpk": "^1.17.1",
     "@types/uuid": "^8.3.4",
     "copyfiles": "^2.4.1",
     "jest": "^27.5.1",
